erms-1347

+++ app/grails-app/controllers/com/k_int/kbplus/UserController.groovy
@@ -34,10 +34,22 @@ class UserController extends AbstractDebugController {
     def _delete() {
         def result = setResultGenerics()
 
+        List<Org> affils = Org.executeQuery('select distinct uo.org from UserOrg uo where uo.user = :user and uo.status = :status',
+                [user: User.get(params.id), status: UserOrg.STATUS_APPROVED])
+
+        if (affils.size() > 1) {
+            flash.error = 'Dieser Nutzer ist mehreren Organisationen zugeordnet und kann daher nicht gelöscht werden.'
+            redirect action: 'edit', params: [id: params.id]
+            return
+        }
+        else if (affils.size() == 1 && (affils.get(0).id != contextService.getOrg().id)) {
+            flash.error = 'Dieser Nutzer ist nicht ihrer Organisationen zugeordnet und kann daher nicht gelöscht werden.'
+            redirect action: 'edit', params: [id: params.id]
+            return
+        }
+
         if (params.process && result.editable) {
-            println params
             User userReplacement = genericOIDService.resolveOID(params.userReplacement)
-            println userReplacement
 
             result.result = deletionService.deleteUser(result.user, userReplacement, false)
         }
